Explicitly declare a string that undergoes mutation is mutable
With Ruby's new immutable string literal mode enabled, treetop fails to parse because it is attempting to mutate a string literal. By explicitly declaring this specific string to be mutable we maintain the same behaviour regardless of whether Ruby's literal strings are all immutable or not.

Note, there may be other places in the codebase where this switch is needed. This is the only place I noticed and things are working ok for me after this change.